+++ mod-pubsub-server/src/main/java/org/folio/dao/MessagingModuleDao.java
@@ -5,7 +5,6 @@ import org.folio.dao.impl.MessagingModuleFilter;
 import org.folio.rest.jaxrs.model.MessagingModule;
 
 import java.util.List;
-import java.util.Optional;
 
 /**
  * Messaging module data access object
@@ -21,30 +20,14 @@ public interface MessagingModuleDao {
   Future<List<MessagingModule>> get(MessagingModuleFilter filter);
 
   /**
-   * Searches {@link MessagingModule} by id
-   *
-   * @param id MessagingModule id
-   * @return future with optional of MessagingModule
-   */
-  Future<Optional<MessagingModule>> getById(String id);
-
-  /**
    * Saves in transaction list of {@link MessagingModule} to data base with specified moduleName
+   * Deletes previous messaging modules with same module name, event type and tenant id as in specified messagingModules
    *
    * @param moduleName  module name
    * @param messagingModules list of Messaging Module entities
    * @return future with list of created Messaging Module entities
    */
   Future<List<MessagingModule>> save(String moduleName, List<MessagingModule> messagingModules);
-
-  /**
-   * Updates {@link MessagingModule} in data base
-   *
-   * @param id messagingModule id
-   * @param messagingModule entity to update
-   * @return future with updated MessagingModule
-   */
-  Future<MessagingModule> update(String id, MessagingModule messagingModule);
 
   /**
    * Deletes {@link MessagingModule} by id