+++ mod-pubsub-server/src/main/java/org/folio/dao/impl/ModuleDaoImpl.java
@@ -9,12 +9,9 @@ import io.vertx.core.logging.LoggerFactory;
 import io.vertx.ext.sql.ResultSet;
 import io.vertx.ext.sql.SQLConnection;
 import io.vertx.ext.sql.UpdateResult;
-import javassist.NotFoundException;
 import org.folio.dao.ModuleDao;
 import org.folio.dao.PostgresClientFactory;
-import org.folio.dao.util.DbUtil;
 import org.folio.rest.jaxrs.model.Module;
-import org.folio.rest.persist.PostgresClient;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
@@ -37,10 +34,8 @@ public class ModuleDaoImpl implements ModuleDao {
   private static final String TABLE_NAME = "module";
   private static final String MODULE_SCHEMA = "pubsub_config";
   private static final String GET_ALL_SQL = "SELECT * FROM %s.%s";
-  private static final String GET_BY_ID_SQL = "SELECT * FROM %s.%s WHERE id = ?";
   private static final String GET_BY_NAME_SQL = "SELECT * FROM %s.%s WHERE name = ?";
   private static final String INSERT_SQL = "INSERT INTO %s.%s (id, name) VALUES (?, ?)";
-  private static final String UPDATE_BY_ID_SQL = "UPDATE %s.%s SET name = ? WHERE id = ?";
   private static final String DELETE_BY_ID_SQL = "DELETE FROM %s.%s WHERE id = ?";
 
   @Autowired
@@ -54,16 +49,6 @@ public class ModuleDaoImpl implements ModuleDao {
     return future.map(this::mapResultSetToModuleList);
   }
 
-  @Override
-  public Future<Optional<Module>> getById(String id) {
-    Future<ResultSet> future = Future.future();
-    String preparedQuery = format(GET_BY_ID_SQL, MODULE_SCHEMA, TABLE_NAME);
-    JsonArray params = new JsonArray().add(id);
-    pgClientFactory.getInstance().select(preparedQuery, params, future.completer());
-    return future.map(resultSet -> resultSet.getResults().isEmpty()
-      ? Optional.empty() : Optional.of(mapRowJsonToModule(resultSet.getRows().get(0))));
-  }
-
   public Future<Optional<Module>> getByName(String name, AsyncResult<SQLConnection> sqlConnection) {
     Future<ResultSet> future = Future.future();
     String preparedQuery = format(GET_BY_NAME_SQL, MODULE_SCHEMA, TABLE_NAME);
@@ -71,12 +56,6 @@ public class ModuleDaoImpl implements ModuleDao {
     pgClientFactory.getInstance().select(sqlConnection, preparedQuery, params, future.completer());
     return future.map(resultSet -> resultSet.getResults().isEmpty()
       ? Optional.empty() : Optional.of(mapRowJsonToModule(resultSet.getRows().get(0))));
-  }
-
-  @Override
-  public Future<String> save(Module module) {
-    PostgresClient pgClient = pgClientFactory.getInstance();
-    return DbUtil.executeInTransaction(pgClient, connection -> save(module, connection));
   }
 
   @Override
@@ -93,23 +72,6 @@ public class ModuleDaoImpl implements ModuleDao {
       future.fail(e);
     }
     return future.map(updateResult -> module.getId());
-  }
-
-  @Override
-  public Future<Module> update(String id, Module module) {
-    Future<UpdateResult> future = Future.future();
-    try {
-      String query = format(UPDATE_BY_ID_SQL, MODULE_SCHEMA, TABLE_NAME);
-      JsonArray params = new JsonArray()
-        .add(module.getName())
-        .add(id);
-      pgClientFactory.getInstance().execute(query, params, future.completer());
-    } catch (Exception e) {
-      LOGGER.error("Error updating Module by id '{}'", e, id);
-      future.fail(e);
-    }
-    return future.compose(updateResult -> updateResult.getUpdated() == 1 ? Future.succeededFuture(module)
-      : Future.failedFuture(new NotFoundException(format("MessagingModule by id '%s' was not updated", id))));
   }
 
   @Override