+++ NEWS.md
@@ -17,6 +17,7 @@
 * Added API for Publishers managing
 * Added API for Subscribers managing
 * Created API for retrieving audit messages
+* Added preliminary cleaning of publisher/subscriber information before declaration publisher/subscriber with same module name and tenant id   
 
  | METHOD |             URL                                                                         | DESCRIPTION                                      |
  |--------|-----------------------------------------------------------------------------------------|--------------------------------------------------|